+++ src/Engine.ml
@@ -741,7 +741,7 @@ module Make (T : TABLE) = struct
   (* We can make this a function of states, or a function of environments. For
      now, the latter appears simpler. *)
 
-  let has_default_reduction env : bool =
+  let env_has_default_reduction env : bool =
     T.default_reduction
       env.current
       (fun _env _prod -> true)